mcp-server-dealer/pyproject.toml
[project]
name = "mcp-server-dealer"
version = "0.1.0"
description = "自動車販売店基幹システム MCPサーバー"
requires-python = ">=3.11"
dependencies = [
    "azure-functions>=1.17.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
package = false
